requirements.txt
python-docx>=1.1.0
PyMuPDF>=1.24.0
groq>=0.9.0
python-dotenv>=1.0.0
rich>=13.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
